Relocated `UmbTiptapToolbarValue`
