+++ src/packages/tiptap/components/types.ts
@@ -1,0 +1,1 @@
+export type UmbTiptapToolbarValue = Array<Array<Array<string>>>;
